model/requirements.txt
aia-auth-client==0.0.8
requests>=2.31.0,<3
httpx>=0.27.2,<1
python-dotenv>=1.0.0,<2
openai>=1.45.0
cachetools>=4.2.2,<6
